reverse builtin <-> runtime dependency
