+++ runtime/Cargo.toml
@@ -33,6 +33,7 @@ serde = { workspace = true }
 serde_repr = { workspace = true }
 thiserror = { workspace = true }
 unsigned-varint = { workspace = true }
+vm_api = { workspace = true } 
 
 # A fake-proofs dependency but... we can't select on that feature here because we enable it from
 # build.rs.
