+++ vm_api/Cargo.toml
@@ -11,14 +11,14 @@ publish = false
 [lib]
 
 [dependencies]
-fil_actors_runtime = { workspace = true, features = [ "test_utils" ] }
-
 anyhow = { workspace = true }
 cid = { workspace = true }
 fvm_ipld_blockstore = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 fvm_ipld_hamt = { workspace = true }
 fvm_shared = { workspace = true }
+num-derive = { workspace = true }
+num-traits = { workspace = true }
 rand = { workspace = true }
 rand_chacha = { workspace = true }
 serde = { workspace = true }
