+++ vm_api/src/lib.rs
@@ -1,23 +1,32 @@
 use std::collections::BTreeMap;
 
 use cid::Cid;
-// TODO: drop the dependency on fil_actors_runtime and have a suitable replacement abstraction here
-// https://github.com/filecoin-project/builtin-actors/issues/1344
-pub use fil_actors_runtime::runtime::{builtins::Type, Primitives};
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_encoding::{
     ipld_block::IpldBlock,
     tuple::{serde_tuple, Deserialize_tuple, Serialize_tuple},
 };
 use fvm_shared::{
-    address::Address, clock::ChainEpoch, econ::TokenAmount, error::ExitCode, MethodNum,
+    address::Address,
+    clock::ChainEpoch,
+    crypto::{
+        hash::SupportedHashes,
+        signature::{Signature, SECP_PUB_LEN, SECP_SIG_LEN, SECP_SIG_MESSAGE_HASH_SIZE},
+    },
+    econ::TokenAmount,
+    error::ExitCode,
+    piece::PieceInfo,
+    sector::RegisteredSealProof,
+    MethodNum,
 };
 
-mod error;
+use builtin::*;
 pub use error::*;
-
-pub mod trace;
 use trace::*;
+
+pub mod builtin;
+mod error;
+pub mod trace;
 pub mod util;
 
 /// An abstract VM that is injected into integration tests
@@ -120,4 +129,37 @@ pub fn new_actor(
         balance,
         delegated_address: predictable_address,
     }
+}
+
+/// Pure functions implemented as primitives by the runtime.
+pub trait Primitives {
+    /// Hashes input data using blake2b with 256 bit output.
+    fn hash_blake2b(&self, data: &[u8]) -> [u8; 32];
+
+    /// Hashes input data using a supported hash function.
+    fn hash(&self, hasher: SupportedHashes, data: &[u8]) -> Vec<u8>;
+
+    /// Hashes input into a 64 byte buffer
+    fn hash_64(&self, hasher: SupportedHashes, data: &[u8]) -> ([u8; 64], usize);
+
+    /// Computes an unsealed sector CID (CommD) from its constituent piece CIDs (CommPs) and sizes.
+    fn compute_unsealed_sector_cid(
+        &self,
+        proof_type: RegisteredSealProof,
+        pieces: &[PieceInfo],
+    ) -> Result<Cid, anyhow::Error>;
+
+    /// Verifies that a signature is valid for an address and plaintext.
+    fn verify_signature(
+        &self,
+        signature: &Signature,
+        signer: &Address,
+        plaintext: &[u8],
+    ) -> Result<(), anyhow::Error>;
+
+    fn recover_secp_public_key(
+        &self,
+        hash: &[u8; SECP_SIG_MESSAGE_HASH_SIZE],
+        signature: &[u8; SECP_SIG_LEN],
+    ) -> Result<[u8; SECP_PUB_LEN], anyhow::Error>;
 }
